Refactor dynamo item and move state to separate file. Add additional functions to domain repository to test the look and feel of dynamo and go.
