+++ lib/adapter/dynamo/item.go
@@ -3,51 +3,13 @@ package dynamo
 import (
 	"github.com/google/uuid"
 	"github.com/grantlerduck/go-was-lambda-dyanmo/lib/domain/booking"
-	"strings"
 )
-
-type State string
 
 const (
-	Unconfirmed     State = "booking-unconfirmed"
-	Confirmed       State = "booking-confirmed"
-	PaymentReceived State = "booking-fee-payed"
-	PaymentPending  State = "booking-fee-pending"
-	Planned         State = "booking-planned"
-	Canceled        State = "booking-canceled"
-	CheckedIn       State = "checked-in"
-	CheckedOut      State = "checked-out"
-	ReviewPending   State = "review-pending"
-	Reviewed        State = "customer-reviewed"
-	Unknown         State = "unknown"
+	HasKeyAttribute  string = "event_id"
+	SortKeyAttribute string = "booking_id"
+	Gsi1KeyAttribute string = "booking_id"
 )
-
-func (s State) String() string {
-	return string(s)
-}
-
-var states = map[State]struct{}{
-	Unconfirmed:     {},
-	Confirmed:       {},
-	PaymentReceived: {},
-	PaymentPending:  {},
-	Planned:         {},
-	Canceled:        {},
-	CheckedIn:       {},
-	CheckedOut:      {},
-	ReviewPending:   {},
-	Reviewed:        {},
-	Unknown:         {},
-}
-
-func getState(str string) State {
-	state := State(strings.ToLower(str))
-	_, ok := states[state]
-	if !ok {
-		return Unknown
-	}
-	return state
-}
 
 type Item struct {
 	EventId     string `json:"event_id" dynamodbav:"event_id"`
@@ -75,4 +37,18 @@ func FromDomainBooking(domain *booking.Event) *Item {
 	item.AirlineName = domain.AirlineName
 	item.State = getState(domain.BookingState)
 	return item
+}
+
+func ToBookingDomain(item Item) *booking.Event {
+	event := new(booking.Event)
+	event.BookingId = item.BookingId
+	event.UserId = item.UserId
+	event.TripFrom = item.TripFrom
+	event.TripUntil = item.TripUntil
+	event.HotelName = item.HotelName
+	event.HotelId = item.HotelId
+	event.FlightId = item.FlightId
+	event.AirlineName = item.AirlineName
+	event.BookingState = item.State.String()
+	return event
 }
